chore: add ticker table component and UI workflow
